Improve gateway-common mock and benchmarking
