+++ xpallets/gateway/common/Cargo.toml
@@ -6,7 +6,7 @@ edition = "2018"
 
 [dependencies]
 codec = { package = "parity-scale-codec", version = "1.3.4", features = ["derive"], default-features = false }
-hex = { version = "0.4", optional = true }
+hex = { version = "0.4", default-features = false, optional = true }
 serde = { version = "1.0", optional = true }
 
 # Substrate primitives
@@ -38,6 +38,7 @@ light-bitcoin = { git = "https://github.com/chainx-org/light-bitcoin", branch = 
 [dev-dependencies]
 lazy_static = "1.4"
 sp-io = "2.0.0"
+frame-benchmarking = "2.0.0"
 pallet-balances = "2.0.0"
 xp-gateway-bitcoin =  { path = "../../../primitives/gateway/bitcoin" }
 xpallet-gateway-bitcoin = { path = "../bitcoin" }
@@ -46,7 +47,7 @@ xpallet-gateway-bitcoin = { path = "../bitcoin" }
 default = ["std"]
 std = [
     "codec/std",
-    "hex",
+    "hex/std",
     "serde",
     # Substrate primitives
     "sp-core/std",
@@ -70,6 +71,7 @@ std = [
     "light-bitcoin/std",
 ]
 runtime-benchmarks = [
+    "hex",
     "sp-core",
     "frame-benchmarking",
 ]
