+++ xpallets/gateway/common/src/tests.rs
@@ -1,8 +1,6 @@
 // Copyright 2019-2020 ChainX Project Authors. Licensed under GPL-3.0.
 
-mod mock;
-
-use mock::*;
+use crate::mock::ExtBuilder;
 
 #[test]
 fn base() {
